+++ prepare-papers.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
-perl /usr/share/texlive/texmf-dist/scripts/pax/pdfannotextractor.pl LNI-Startseiten
+PAXDIR=`kpsewhich -var-value TEXMFDIST`/scripts/pax
+
+perl $PAXDIR/pdfannotextractor.pl LNI-Startseiten
 cd papers/
 for d in */ ; do
 	cd "$d"
 	echo "Processing $d.." 
-	perl /usr/share/texlive/texmf-dist/scripts/pax/pdfannotextractor.pl paper
+	perl $PAXDIR/pdfannotextractor.pl paper
 	echo "Finished with $d"
 	cd ..
 done